Release of `sesssion` in beta state
