+++ deno.json
@@ -1,6 +1,6 @@
 {
   "name": "@vixeny/core",
-  "version": "0.1.46",
+  "version": "0.1.47",
   "exports": "./main.ts",
   "imports": {
     "@cross/test": "jsr:@cross/test@^0.0.9",
